requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
requests